Extracting classes from AgendaService.java
